+++ tornapi/endpoints/__init__.py
@@ -1,3 +1,4 @@
 from .user import UserEndpoint
+from .torn import TornEndpoint
 
-__all__ = ['UserEndpoint']
+__all__ = ['UserEndpoint', 'TornEndpoint']